telegram-bot/requirements.txt
# Основные зависимости
aiogram==3.24.0
aiohttp==3.9.0
pydantic==2.7.0
pydantic-settings==2.12.0
python-dotenv==1.0.0
redis==5.0.1

# Для очереди сообщений
aiofiles==23.2.1

# Разработка и тестирование
black==23.11.0
isort==5.12.0
mypy==1.7.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Типизация
types-aiofiles==23.2.0.20240623
types-redis==4.6.0.10